+++ sdk/storage/storage-queue/test/messageidclient.spec.ts
@@ -1,10 +1,7 @@
 import * as assert from "assert";
-import { MessageIdClient } from '../src';
 import { getQSU, sleep } from "./utils";
 import { QueueClient } from "../src/QueueClient";
 import { record } from "./utils/recorder";
-import { TokenCredential } from '@azure/core-http';
-import { assertClientUsesTokenCredential } from './utils/assert';
 import * as dotenv from "dotenv";
 dotenv.config({ path: "../.env" });
 
@@ -162,16 +159,5 @@ describe("MessageIdClient", () => {
       error = err;
     }
     assert.ok(error);
-  });
-
-  it("can be created with a url and a TokenCredential", async () => {
-    const tokenCredential: TokenCredential = {
-      getToken: () => Promise.resolve({
-        token: 'token',
-        expiresOnTimestamp: 12345
-      })
-    }
-    const newClient = new MessageIdClient("https://queue", tokenCredential);
-    assertClientUsesTokenCredential(newClient);
   });
 });
