+++ sdk/storage/storage-queue/test/messagesclient.spec.ts
@@ -1,10 +1,7 @@
 import * as assert from "assert";
-import { MessagesClient } from '../src';
 import { getQSU } from "./utils";
 import { QueueClient } from "../src/QueueClient";
 import { record } from "./utils/recorder";
-import { TokenCredential } from '@azure/core-http';
-import { assertClientUsesTokenCredential } from './utils/assert';
 import * as dotenv from "dotenv";
 dotenv.config({ path: "../.env" });
 
@@ -343,16 +340,5 @@ describe("MessagesClient", () => {
         "The request body is too large and exceeds the maximum permissible limit."
       )
     );
-  });
-
-  it("can be created with a url and a TokenCredential", async () => {
-    const tokenCredential: TokenCredential = {
-      getToken: () => Promise.resolve({
-        token: 'token',
-        expiresOnTimestamp: 12345
-      })
-    }
-    const newClient = new MessagesClient("https://queue", tokenCredential);
-    assertClientUsesTokenCredential(newClient);
   });
 });
