+++ sdk/storage/storage-queue/test/node/messageidclient.spec.ts
@@ -1,9 +1,12 @@
 import * as assert from "assert";
+import { MessageIdClient } from '../../src';
 import { getQSU, getConnectionStringFromEnvironment } from "../utils";
 import { record } from "../utils/recorder";
 import { QueueClient } from "../../src/QueueClient";
 import { MessagesClient } from "../../src/MessagesClient";
 import { SharedKeyCredential } from "../../src/credentials/SharedKeyCredential";
+import { TokenCredential } from '@azure/core-http';
+import { assertClientUsesTokenCredential } from '../utils/assert';
 
 describe("MessageIdClient Node.js only", () => {
   const queueServiceClient = getQSU();
@@ -13,7 +16,7 @@ describe("MessageIdClient Node.js only", () => {
 
   let recorder: any;
 
-  beforeEach(async function() {
+  beforeEach(async function () {
     recorder = record(this);
     queueName = recorder.getUniqueName("queue");
     queueClient = queueServiceClient.createQueueClient(queueName);
@@ -179,5 +182,16 @@ describe("MessageIdClient Node.js only", () => {
         "Error message is different than expected."
       );
     }
+  });
+
+  it("can be created with a url and a TokenCredential", async () => {
+    const tokenCredential: TokenCredential = {
+      getToken: () => Promise.resolve({
+        token: 'token',
+        expiresOnTimestamp: 12345
+      })
+    }
+    const newClient = new MessageIdClient("https://queue", tokenCredential);
+    assertClientUsesTokenCredential(newClient);
   });
 });
