+++ sdk/storage/storage-queue/test/node/messagesclient.spec.ts
@@ -4,6 +4,8 @@ import { record } from "../utils/recorder";
 import { QueueClient } from "../../src/QueueClient";
 import { MessagesClient } from "../../src/MessagesClient";
 import { SharedKeyCredential } from "../../src/credentials/SharedKeyCredential";
+import { TokenCredential } from '@azure/core-http';
+import { assertClientUsesTokenCredential } from '../utils/assert';
 
 describe("MessagesClient Node.js only", () => {
   const queueServiceClient = getQSU();
@@ -13,7 +15,7 @@ describe("MessagesClient Node.js only", () => {
 
   let recorder: any;
 
-  beforeEach(async function() {
+  beforeEach(async function () {
     recorder = record(this);
     queueName = recorder.getUniqueName("queue");
     queueClient = queueServiceClient.createQueueClient(queueName);
@@ -187,5 +189,16 @@ describe("MessagesClient Node.js only", () => {
         "Error message is different than expected."
       );
     }
+  });
+
+  it("can be created with a url and a TokenCredential", async () => {
+    const tokenCredential: TokenCredential = {
+      getToken: () => Promise.resolve({
+        token: 'token',
+        expiresOnTimestamp: 12345
+      })
+    }
+    const newClient = new MessagesClient("https://queue", tokenCredential);
+    assertClientUsesTokenCredential(newClient);
   });
 });
