+++ sdk/storage/storage-queue/test/node/queueserviceclient.spec.ts
@@ -4,11 +4,13 @@ import { record } from "../utils/recorder";
 import { QueueServiceClient } from "../../src/QueueServiceClient";
 import { SharedKeyCredential } from "../../src/credentials/SharedKeyCredential";
 import { newPipeline } from "../../src";
+import { TokenCredential } from '@azure/core-http';
+import { assertClientUsesTokenCredential } from '../utils/assert';
 
 describe("QueueServiceClient Node.js only", () => {
   let recorder: any;
 
-  beforeEach(function() {
+  beforeEach(function () {
     recorder = record(this);
   });
 
@@ -70,5 +72,16 @@ describe("QueueServiceClient Node.js only", () => {
 
     assert.ok(typeof result.requestId);
     assert.ok(result.requestId!.length > 0);
+  });
+
+  it("can be created with a url and a TokenCredential", async () => {
+    const tokenCredential: TokenCredential = {
+      getToken: () => Promise.resolve({
+        token: 'token',
+        expiresOnTimestamp: 12345
+      })
+    }
+    const newClient = new QueueServiceClient("https://queue", tokenCredential);
+    assertClientUsesTokenCredential(newClient);
   });
 });
